scrapers/requirements.txt
# Core Scrapy framework
scrapy>=2.11.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
SQLAlchemy>=2.0.23

# Redis for state management
redis>=5.0.1

# HTTP and networking
requests>=2.31.0
httpx>=0.25.0
urllib3>=2.1.0

# Playwright for JavaScript rendering
playwright>=1.40.0
scrapy-playwright>=0.0.33

# Data processing
beautifulsoup4>=4.12.2
lxml>=4.9.3
PyYAML>=6.0.1

# JSON and data validation
jsonschema>=4.20.0
pydantic>=2.5.2
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Observability (OpenTelemetry)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation>=0.42b0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-psycopg2>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-asyncpg>=0.42b0
structlog>=23.2.0

# Logging and monitoring
loguru>=0.7.2
prometheus-client>=0.19.0

# Reddit API
praw>=7.7.1

# Music service APIs
tidalapi>=0.8.0

# Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
